backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-dotenv==1.0.1
psycopg2-binary==2.9.11
pillow==12.1.0
openpyxl==3.1.5
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.1
django-cors-headers==4.3.1
gunicorn==22.0.0
opencv-python-headless==4.10.0.84
torch
torchvision
face_recognition
pandas
